just what i needed to bust. thank you based god

+++ src/wnd/wincludes.h
@@ -55,7 +55,7 @@ extern bool m_base_open[ OBJ_BASE_COUNT / 2 ];
 extern bool menu_style_toggle[ OBJ_BASE_COUNT / 2 ];
 
 extern bool macro_recording;
-extern std::vector< u32 > macro;
+extern std::vector< u8 > macro;
 
 enum MENU_STYLE {
   FPS = 0,
